% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbitmap.R
\name{ggbitmap}
\alias{ggbitmap}
\title{Transform a tiff plot to a ggplot2 object, that can be used by patchwork package.}
\usage{
ggbitmap(bitmap)
}
\arguments{
\item{bitmap}{a tiff picture}
}
\value{
a ggplot2 object
}
\description{
Transform a tiff plot to a ggplot2 object, that can be used by patchwork package.
}
\examples{
ggbitmap("venn.tiff")
}
\author{
ZhonghuiGai
}
