% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvenn.R
\name{ggvenn}
\alias{ggvenn}
\title{Venn plot based on ggplot2, forked from yanlinlin82/ggvenn}
\usage{
ggvenn(data, text.size = 4, label.color = "white")
}
\arguments{
\item{text.size}{the size of label and text size, default value is 4}
}
\value{

}
\description{
Venn plot based on ggplot2, forked from yanlinlin82/ggvenn
}
\examples{
data <- read.csv(file = "genus.Syn.csv", row.names = 1)
data <- ggvennEx:::vennlist(data)
v.d <- venn.data(data)
ggvenn(v.d)
}
\author{
ZHonghui Gai
}
