% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvennplot.R
\name{ggvennplot}
\alias{ggvennplot}
\title{Draw a venn plot using VennDiagram and transform it to ggplot2 object}
\usage{
ggvennplot(data, col = c("#5dcca0", "#ff445d"), delete = TRUE)
}
\arguments{
\item{data}{a data frame containing grouping information}

\item{col}{colour used to draw venn plot, the length is equal to the levels of group}

\item{delete}{boolean value to delete the original .tiff file, default value id TRUE}
}
\value{

}
\description{
Draw a venn plot using VennDiagram and transform it to ggplot2 object
}
\examples{
data <- read.csv(file = "genus.Syn.csv", row.names = 1)
#data <- ggvennEx:::vennlist(data)
ggvennplot(data = data, col = c("red", "green", "blue", "orange"))
ggvennplot(data = data[data$group != "3W", ], col = c("red", "blue", "orange"))
}
\author{
Zhonghui Gai
}
