% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/son.plot.R
\name{son.plot}
\alias{son.plot}
\title{Insert a plot to another plot using ggplot2}
\usage{
son.plot(mother.plot, son.plot, xmin, xmax, ymin, ymax, venn = FALSE)
}
\arguments{
\item{mother.plot}{The plot which to be inserted}

\item{son.plot}{The plot to be inserted into other plot}

\item{xmin}{the position of the son.plot}

\item{xmax}{the position of the son.plot}

\item{ymin}{the position of the son.plot}

\item{ymax}{the position of the son.plot}

\item{venn}{boolean value}
}
\value{

}
\description{
Insert a plot to another plot using ggplot2
}
\examples{
p1 <- ggplot(data = iris, aes(x = Sepal.Length, y = Sepal.Width)) +
geom_point() + theme_transparent()
p2 <- ggvennplot(data = data[data$group != "3W", ], col = c("red", "blue", "orange"))
son.plot(p2, p1, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
}
\author{
Zhonghui Gai
}
