% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_transparent.R
\name{theme_transparent}
\alias{theme_transparent}
\title{A transparent layer function}
\usage{
theme_transparent(...)
}
\arguments{
\item{...}{other paramaters}
}
\value{

}
\description{
A transparent layer function
}
\examples{
theme_transparent()
}
\author{
Zhonghui Gai
}
