% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.data.R
\name{venn.data}
\alias{venn.data}
\title{Generate input data for ggvenn()}
\usage{
venn.data(data)
}
\arguments{
\item{data}{input data, must be a list, this can be generated form vennlist() function}
}
\value{
a list as the input of ggvenn() function
}
\description{
Generate input data for ggvenn()
}
\examples{
data <- read.csv(file = "genus.Syn.csv", row.names = 1)
data <- ggvennEx:::vennlist(data)
v.d <- venn.data(data)
}
\author{
ZHonghui Gai
}
